chore: move json parser to directory and fix any type
